Create new module to switch screens
